requirements.txt
streamlit>=1.24.0
openai>=0.27.0
requests>=2.28.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
colorama>=0.4.6
pytest>=7.0.0
